# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API E-commerce Segura con Swagger Completo",
        "description": "API RESTful para la gestión de un e-commerce: registro y autenticación de clientes mediante tokens, catálogo de productos, creación de órdenes de compra y procesamiento de pagos con Stripe.",
        "contact": {
            "email": "soporte@ecommerce-api.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Órdenes"
                ],
                "summary": "Historial de compras del usuario autenticado",
                "description": "Devuelve las órdenes del cliente autenticado, ordenadas de la más reciente a la más antigua.",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "Listado paginado de órdenes del cliente"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Órdenes"
                ],
                "summary": "Crear una orden de compra",
                "description": "Crea una orden a partir de una lista de productos y cantidades. Valida el stock disponible, descuenta el inventario, guarda una fotografía (snapshot) del nombre y precio de cada producto, y crea un PaymentIntent en Stripe por el total de la orden. La respuesta incluye el \"client_secret\" necesario para confirmar el pago desde un frontend, o puede confirmarse directamente desde el endpoint POST /api/orders/{order}/confirm-payment.",
                "operationId": "2123ba58dadb22bd6f2c8081af562230",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Orden creada exitosamente, incluye client_secret de Stripe"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Error de validación o stock insuficiente"
                    },
                    "502": {
                        "description": "Error al comunicarse con Stripe"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{order}": {
            "get": {
                "tags": [
                    "Órdenes"
                ],
                "summary": "Detalle de una orden propia",
                "description": "Devuelve el detalle completo (ítems y pago) de una orden que pertenezca al cliente autenticado.",
                "operationId": "8114ac185cc1bf301d8862617e497098",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de la orden"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "La orden pertenece a otro cliente"
                    },
                    "404": {
                        "description": "Orden no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{order}/confirm-payment": {
            "post": {
                "tags": [
                    "Órdenes",
                    "Pagos"
                ],
                "summary": "Confirmar o sincronizar el pago de una orden",
                "description": "Si se envía \"payment_method\" (ej. un token de prueba de Stripe como \"pm_card_visa\"), confirma el PaymentIntent directamente desde la API — útil para probar el flujo completo desde Swagger UI sin un frontend. Si se omite, simplemente sincroniza el estado local con el estado actual del PaymentIntent en Stripe (útil cuando la confirmación ya ocurrió en el cliente vía Stripe.js).",
                "operationId": "dda8f9cc2370e626ba92e67b8fa267ba",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "payment_method": {
                                        "description": "Token de método de pago de prueba de Stripe. Opcional.",
                                        "type": "string",
                                        "example": "pm_card_visa"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pago confirmado o sincronizado exitosamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "402": {
                        "description": "El pago fue rechazado por Stripe"
                    },
                    "403": {
                        "description": "La orden pertenece a otro cliente"
                    },
                    "404": {
                        "description": "Orden o pago no encontrado"
                    },
                    "422": {
                        "description": "La orden ya fue procesada previamente"
                    },
                    "502": {
                        "description": "Error al comunicarse con Stripe"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtenido en /api/register o /api/login. Enviar como: Authorization: Bearer {token}",
                "bearerFormat": "Sanctum Personal Access Token",
                "scheme": "bearer"
            }
        }
    },
...
}